add more folders to individual and create safe_mv
I found out hat using mv with identical source and
target breaks it so I decided to wrap it in safe_mv
although the failed mv command was not a problem
since the script continued but this is just for
completeness sake. I guess I never realised that
because I never executed this script more than once
a day, in fact many times much more rarely.