Fix error within update function
Even though I created a safe_mv function which took into consideration
source and target being equal it still did not work with wildcards
though they appeared to be resolved somehow. Nevertheless the following
error still occured:
mv: cannot move '...A-2024-04-06/' to a subdirectory of itself, '...A-2024-04-06/A-2024-04-06'
Adding realpath on the source argument appears to have resolved the
issue.